pyproject.toml
[project]
name = "nebula-prime-adk"
version = "0.1.0"
description = "MultiAgentic AI using ADK to help reduce clinician burnout."
authors = [
    {name = "Atiur Siddique", email = "atiurrs@gmail.com"},
    {name = "Kamrul Islam", email = "kislam@gmail.com"},
    {name = "Matiur Siddiqui", email = "siddiquimatiur@gmail.com"}
]
license = "Apache License 2.0"
readme = "README.md"

dependencies = [
    "google-adk(==1.3.0)",
    "google-api-core(==2.25.0)",
    "google-api-python-client(==2.171.0)",
    "google-auth(==2.40.3)",
    "google-auth-httplib2(==0.2.0)",
    "google-cloud-aiplatform(==1.97.0)",
    "google-cloud-appengine-logging(==1.6.1)",
    "google-cloud-audit-log(==0.3.2)",
    "google-cloud-bigquery(==3.34.0)",
    "google-cloud-core(==2.4.3)",
    "google-cloud-logging(==3.12.1)",
    "google-cloud-resource-manager(==1.14.2)",
    "google-cloud-secret-manager(==2.24.0)",
    "google-cloud-speech(==2.32.0)",
    "google-cloud-storage(==2.19.0)",
    "google-cloud-trace(==1.16.1)",
    "google-crc32c(==1.7.1)",
    "google-genai(==1.19.0)",
    "google-resumable-media(==2.7.2)",
    "googleapis-common-protos(==1.70.0)",
    "grpc-google-iam-v1(==0.14.2)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-dotenv (==1.1.0)",
    "pydantic(==2.11.5)",
    "pydantic-settings(==2.9.1)",
    "pydantic_core(==2.33.2)",
    "litellm (==1.72.4)",
    "absl-py (==2.3.0)",
    "cloudpickle (==3.1.1)",
]

[tool.poetry.dependencies]
google-cloud-aiplatform = { version = "^1.93.0", extras = [
    "adk",
    "agent-engines",
] }
python = "^3.9"
google-genai = "^1.9.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
google-adk = "^1.0.0"
litellm = "^1.72.4"
[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
google-adk = { version = "^1.0.0", extras = ["eval"] }
pytest-asyncio = "^0.26.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "deployment"

[tool.poetry.scripts]
depoy-agent-engine = "deployment.deploy:main"
